requirements.txt
# Function dependencies
google-cloud-storage==2.*
supabase==1.*